Develop

+++ app/public/js/status-auth.js
@@ -1,10 +1,39 @@
-// AUTH STATUS CHANGES LISTENER
-auth.onAuthStateChanged(user => {
-    if(user) {
-        console.log('user logged in: ',user);
-    } else {
-        console.log('user logged out');
-        localStorage.clear();          //LOCALSTORAGE REMOVE
-        window.location = '/login';
-    }
-});
+var clientId = "990666211388-cb76b22m9gnvn7e8b99mpkc2ptp8vp37.apps.googleusercontent.com";
+
+function gapiInit() {                   
+
+    gapi.load("auth2", function () {
+
+        var googleOauthClient;
+
+        gapi.auth2.init({
+            client_id: clientId
+        }).then((client) => {
+
+            googleOauthClient = gapi.auth2.getAuthInstance();
+        
+
+            if (googleOauthClient.isSignedIn.get() == false) { 
+
+                // AUTH STATUS CHANGES LISTENER
+                auth.onAuthStateChanged(user => {
+                    if(user) {
+                        console.log('user logged in: ',user);
+                    } 
+
+                    else {
+                        console.log('user logged out');
+                        localStorage.clear();          //LOCALSTORAGE REMOVE
+                        window.location = '/login';
+                    }
+                });
+
+            }
+
+        }).catch((err) => {
+            return -1;
+        });
+
+    });
+    
+}
